+++ main.py
@@ -47,6 +47,7 @@ import utils # Import the utils.py scripts.
 style = utils.style # Load the style from the utils script.
 clear = utils.clear # Load the screen clearing function from the utils script.
 prompt = utils.prompt # Load the user input prompt function from the utils script.
+play_sound = utils.play_sound # Load the function used to play sounds from the utils script.
 display_message = utils.display_message # Load the message display function from the utils script.
 process_gpx = utils.process_gpx # Load the GPX processing function from the utils script.
 save_to_file = utils.save_to_file # Load the file saving function from the utils script.
@@ -141,7 +142,6 @@ real_time_bottom_margin = config["realtime"]["real_time_bottom_margin"] # How ma
 real_time_image_rotation = config["realtime"]["real_time_image_rotation"] # How many degrees clockwise the image will be rotated in real-time mode.
 fswebcam_device = config["realtime"]["fswebcam_device"] # This setting determines the video device that 'fswebcam' will use to take images in real-time mode.
 fswebcam_flags = config["realtime"]["fswebcam_flags"] # These are command flags that will be added to the end of the FSWebcam command. You can use these to customize how FSWebcam takes images in real-time mode based on your camera set up.
-audio_alerts = config["realtime"]["audio_alerts"] # This setting determines whether or not Predator will make use of sounds to inform the user of events.
 webhook = config["realtime"]["webhook"] # This setting can be used to define a webhook that Predator will send a request to when it detects a license plate in real-time mode. See CONFIGURATION.md to learn more about how to use flags in this setting.
 shape_alerts = config["realtime"]["shape_alerts"] # This setting determines whether or not prominent text-based shapes will be displayed for various actions. This is useful in vehicle installations where you may want to see whether or not Predator detected a plate at a glance.
 save_real_time_object_recognition = config["realtime"]["save_real_time_object_recognition"] # This setting determines whether or not Predator will save the objects detected in real-time mode to a file. When this is turned off, object recognition data will only be printed to the console.
@@ -162,18 +162,6 @@ gotify_application_token = config["realtime"]["gotify_application_token"] # This
 status_lighting_enabled = config["realtime"]["status_lighting_enabled"]
 status_lighting_base_url = config["realtime"]["status_lighting_base_url"]
 status_lighting_values = config["realtime"]["status_lighting_values"]
-
-# Audio alert settings
-alert_sounds_startup = config["realtime"]["startup_sound"]["path"]
-alert_sounds_startup_repeat = config["realtime"]["startup_sound"]["repeat"]
-alert_sounds_startup_delay = config["realtime"]["startup_sound"]["delay"]
-alert_sounds_notice = config["realtime"]["notification_sound"]["path"]
-alert_sounds_notice_repeat = config["realtime"]["notification_sound"]["repeat"]
-alert_sounds_notice_delay = config["realtime"]["notification_sound"]["delay"]
-alert_sounds_alert = config["realtime"]["alert_sound"]["path"]
-alert_sounds_alert_repeat = config["realtime"]["alert_sound"]["repeat"]
-alert_sounds_alert_delay = config["realtime"]["alert_sound"]["delay"]
-
 
 
 
@@ -214,10 +202,7 @@ else: # If the user his disabled the large ASCII art header, then show a simple 
     if (custom_startup_message != ""): # Only display the line for the custom message if the user has defined one.
         print(custom_startup_message) # Show the user's custom defined start-up message.
 
-if (audio_alerts == True and int(alert_sounds_startup_repeat) > 0): # Check to see if the user has audio alerts enabled.
-    for i in range(0, int(alert_sounds_startup_repeat)): # Repeat the sound several times, if the configuration says to.
-        os.system("mpg321 " + alert_sounds_startup + " > /dev/null 2>&1 &") # Play the sound specified for this alert type in the configuration.
-        time.sleep(float(alert_sounds_startup_delay)) # Wait before playing the sound again.
+play_sound("startup")
 
 if (push_notifications_enabled == True): # Check to see if the user has push notifications enabled.
     os.system("curl -X POST '" + gotify_server + "/message?token=" + gotify_application_token + "' -F 'title=Predator' -F 'message=Predator has been started.' > /dev/null 2>&1 &") # Send a push notification via Gotify indicating that Predator has started.
@@ -1492,10 +1477,7 @@ elif (mode_selection == "2" and realtime_mode_enabled == True): # The user has s
                     if (realtime_output_level >= 2): # Only display this status message if the output level indicates to do so.
                         print("Detected Plate: " + detected_plate) # Print the detected plate.
 
-                    if (audio_alerts == True and int(alert_sounds_notice_repeat) > 0): # Check to see if the user has audio alerts enabled.
-                        for i in range(0, int(alert_sounds_notice_repeat)): # Repeat the sound several times, if the configuration says to.
-                            os.system("mpg321 " + alert_sounds_notice + " > /dev/null 2>&1 &") # Play the sound specified for this alert type in the configuration.
-                            time.sleep(float(alert_sounds_notice_delay)) # Wait before playing the sound again.
+                    play_sound("notification")
 
                     if (push_notifications_enabled == True): # Check to see if the user has Gotify notifications enabled.
                         os.system("curl -X POST '" + gotify_server + "/message?token=" + gotify_application_token + "' -F 'title=Predator' -F 'message=A license plate has been detected: " + detected_plate + "' > /dev/null 2>&1 &") # Send a push notification via Gotify.
@@ -1559,10 +1541,7 @@ elif (mode_selection == "2" and realtime_mode_enabled == True): # The user has s
                                 if (status_lighting_enabled == True): # Check to see if status lighting alerts are enabled in the Predator configuration.
                                     update_status_lighting("alert") # Run the function to update the status lighting.
 
-                                if (audio_alerts == True and int(alert_sounds_alert_repeat) > 0): # Check to see if the user has audio alerts enabled.
-                                    for i in range(0, int(alert_sounds_alert_repeat)): # Repeat the sound several times, if the configuration says to.
-                                        os.system("mpg321 " + alert_sounds_alert + " > /dev/null 2>&1 &") # Play the sound specified for this alert type in the configuration.
-                                        time.sleep(float(alert_sounds_alert_delay)) # Wait before playing the sound again.
+                                play_sound("alert")
 
                                 if (push_notifications_enabled == True): # Check to see if the user has Gotify notifications enabled.
                                     os.system("curl -X POST '" + gotify_server + "/message?token=" + gotify_application_token + "' -F 'title=Predator' -F 'message=A license plate in the alert database has been detected: " + str(individual_detected_plate_guess) + "' > /dev/null 2>&1 &") # Send a push notification using Gotify.
@@ -1593,10 +1572,7 @@ elif (mode_selection == "2" and realtime_mode_enabled == True): # The user has s
                         if (status_lighting_enabled == True): # Check to see if status lighting alerts are enabled in the Predator configuration.
                             update_status_lighting("alert") # Run the function to update the status lighting.
 
-                        if (audio_alerts == True and int(alert_sounds_alert_repeat) > 0): # Check to see if the user has audio alerts enabled.
-                            for i in range(0, int(alert_sounds_alert_repeat)): # Repeat the sound several times, if the configuration says to.
-                                os.system("mpg321 " + alert_sounds_alert + " > /dev/null 2>&1 &") # Play the sound specified for this alert type in the configuration.
-                                time.sleep(float(alert_sounds_alert_delay)) # Wait before playing the sound again.
+                        play_sound("alert")
 
                         if (push_notifications_enabled == True): # Check to see if the user has Gotify notifications enabled.
                             os.system("curl -X POST '" + gotify_server + "/message?token=" + gotify_application_token + "' -F 'title=Predator' -F 'message=A license plate in the alert database has been detected: " + detected_plate + "' > /dev/null 2>&1 &") # Send a push notification using Gotify.